Fix enable/disabling of publish all.
The publish all option was incorrectly using a bool type which will enable publishing of all devices
forever once set due to always being true if it existed in the command arguments and was saved to the config.

This changes the type to int to allow it to be set to 0 to disable it.