The main loop should not run in the same scope as __init__.
This package runs the main application directly from __init__.py.

This approach prevents other applications from being able to import the
modules exposed by the app, which is a bit of a shame because this
package exports a pre-compiled version of the TheengsDecoder and it
would be nice to import it in other packages without running the whole
gateway.

However, without this PR this code would hang:

>>> from TheengsGateway._decoder import decodeBLE as dble

The reason is that `run()` is executed in the static scope of
__init__.py, and therefore the loop is started and joined as soon as the
parent module is imported.

This PR addresses this issue by wrapping the main app logic into a
`main` method, and adding a separate __main__.py, as it's common
practice in Python.